DEVSU-2117 Reorganize Genomic Report Print
- DEVSU-2117
- Split genomic summary component to patient info + tumour summary and key alterations
- Rearrange therapeutic targets and chemoresistance tables in between tumour summary and key alterations
- Reduce cell white space and padding for therapeutic targets and chemoresistance table rows
- Make therapeutic targets and chemoresistance tables full width in print mode, even when empty
