+++ app/views/PrintView/index.tsx
@@ -200,15 +200,21 @@ const Print = ({
     if (report && template) {
       return (
         <>
-          {template?.sections.includes('summary') && (
-            <>
-              <Summary templateName={report.template.name} isPrint printVersion={printVersion} loadedDispatch={dispatch} />
-              <PageBreak />
-            </>
+          {template?.sections.includes('summary') && template?.name !== 'genomic' && (
+            <Summary templateName={report.template.name} isPrint printVersion={printVersion} loadedDispatch={dispatch} />
+          )}
+          {template?.sections.includes('summary') && template?.name === 'genomic' && ( // Splitting the patient info and tumour summary sections from alterations for genomic reports to insert therapeutic targets
+            <Summary templateName="genomicPatientandTumour" isPrint printVersion={printVersion} loadedDispatch={dispatch} />
           )}
           {template?.sections.includes('therapeutic-targets') && (
             <>
               <TherapeuticTargets isPrint loadedDispatch={dispatch} />
+              <PageBreak />
+            </>
+          )}
+          {template?.sections.includes('summary') && template?.name === 'genomic' && ( // Continuing key alterations after therapeutic targets for genomic reports
+            <>
+              <Summary templateName="genomicAlterations" isPrint printVersion={printVersion} />
               <PageBreak />
             </>
           )}
